+++ Plugin/CustomPosters/src/Data/ShipWindows.cs
@@ -4,14 +4,79 @@ namespace CustomPosters.Data.PosterLayouts
 {
     internal static class ShipWindows
     {
-        public static PosterData[] Get() => new PosterData[]
+        public static PosterData[] Get()
         {
-            new() { Position = new Vector3(4.1886f, 2.9318f, -16.8409f), Rotation = new Vector3(0, 200.9872f, 0), Scale = new Vector3(0.6391f, 0.4882f, 2f), Name = "Poster1" },
-            new() { Position = new Vector3(6.4202f, 2.2577f, -10.8226f), Rotation = new Vector3(0, 0, 0), Scale = new Vector3(0.7296f, 0.4896f, 1f), Name = "Poster2" },
-            new() { Position = new Vector3(9.9186f, 2.8591f, -17.4716f), Rotation = new Vector3(0, 180f, 356.3345f), Scale = new Vector3(0.7487f, 1.0539f, 1f), Name = "Poster3" },
-            new() { Position = new Vector3(6.4449f, 3.0961f, -10.8221f), Rotation = new Vector3(0, 0.026f, 2.68f), Scale = new Vector3(0.7289f, 0.9989f, 1f), Name = "Poster4" },
-            new() { Position = new Vector3(5.5286f, 2.5882f, -17.3541f), Rotation = new Vector3(0, 201.1556f, 359.8f), Scale = new Vector3(0.5516f, 0.769f, 1f), Name = "Poster5" },
-            new() { Position = new Vector3(3.0647f, 2.8174f, -11.7341f), Rotation = new Vector3(0, 0, 358.6752f), Scale = new Vector3(0.8596f, 1.2194f, 1f), Name = "CustomTips" }
-        };
+            bool usePoster5Vanilla = Plugin.ModConfig.UsePoster5VanillaModel.Value;
+            bool useTipsVanilla = Plugin.ModConfig.UseTipsVanillaModel.Value;
+
+            return new PosterData[]
+            {
+                new()
+                {
+                    Position = new Vector3(4.1886f, 2.9318f, -16.8409f),
+                    Rotation = new Vector3(0, 200.9872f, 0),
+                    Scale = new Vector3(0.6391f, 0.4882f, 2f),
+                    Name = "Poster1"
+                },
+                new()
+                {
+                    Position = new Vector3(6.4202f, 2.2577f, -10.822f),
+                    Rotation = new Vector3(0, 0, 0),
+                    Scale = new Vector3(0.7296f, 0.4896f, 1f),
+                    Name = "Poster2"
+                },
+                new()
+                {
+                    Position = new Vector3(9.9186f, 2.8591f, -17.4716f),
+                    Rotation = new Vector3(0, 180f, 356.3345f),
+                    Scale = new Vector3(0.7487f, 1.0539f, 1f),
+                    Name = "Poster3"
+                },
+                new()
+                {
+                    Position = new Vector3(6.4449f, 3.0961f, -10.8219f),
+                    Rotation = new Vector3(0, 0, 2.68f),
+                    Scale = new Vector3(0.7289f, 0.9989f, 1f),
+                    Name = "Poster4"
+                },
+
+
+                usePoster5Vanilla
+                    ? new()
+                    {
+                        // Vanilla model
+                        Position = new Vector3(5.5286f, 2.5882f, -17.3421f),
+                        Rotation = new Vector3(1.3609f, 0.2388f, 182.4321f),
+                        Scale = new Vector3(0.465f, 0.71f, 1f),
+                        Name = "Poster5"
+                    }
+                    : new()
+                    {
+                        // Quad
+                        Position = new Vector3(5.5286f, 2.5882f, -17.3541f),
+                        Rotation = new Vector3(0, 201.1556f, 359.8f),
+                        Scale = new Vector3(0.5516f, 0.769f, 1f),
+                        Name = "Poster5"
+                    },
+
+                useTipsVanilla
+                    ? new()
+                    {
+                        // Vanilla model
+                        Position = new Vector3(8.1417f, 2.517138f, -21.9378f),
+                        Rotation = new Vector3(-270.392f, -225.379f, 314.887f),
+                        Scale = new Vector3(46.75954f, 100f, 70.89838f),
+                        Name = "CustomTips"
+                    }
+                    : new()
+                    {
+                        // Quad
+                        Position = new Vector3(3.0647f, 2.8174f, -11.7255f),
+                        Rotation = new Vector3(0, 0, 358.6752f),
+                        Scale = new Vector3(0.8596f, 1.2194f, 1f),
+                        Name = "CustomTips"
+                    }
+            };
+        }
     }
 }