+++ Thunderstore/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2025-10-21 - "Vanilla models"
+### Added
+- Added an option to use Poster5 and Tips vanilla models (default).
+
+### Changed
+- Poster positions and rotation of ShipWindows, TwoStoryShip, and WiderShipMod were improved to be closer to the original.
+
 ## [3.2.0] - 2025-10-15 - "Synchronization"
 ### Added
 - Added video playback synchronization. Hosts video playback time is synced with all clients.